adds 'tables' utility (and specifically 'tables/columns')
